Add initial Postgresql support.
I am not sure what the naming scheme should be for the database tables file.
I just chose to add `.pgsql` to the name.
I left the MySQL one alone.

The DSN for Postgresql needs to be something more like this:
```
dsn = "pgsql:host={{host}};dbname={{database}};user={{user}};password={{password}}"
```

The widgets being returned is an object rather than an array.
This causes a lot of problems with fatal errors.
Cast it to an array.
However, there are likely more changes that is needed when it comes to files.
This widget structure may need to be toggled to be an array somehow.
This might be doable via the PDO or related settings.

Fix some improper conditionally checks, adding `isset()` or similar as needed.